api/requirements.txt
connexion == 2.14.2
connexion[swagger-ui] == 2.14.2
python_dateutil == 2.8.2
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
Flask~=2.2.5
six~=1.17.0
PyYAML~=6.0.2
colorama~=0.4.6
pytest~=8.3.5
Werkzeug~=2.2.3
pymongo>=3.10.0
boto3>=1.26.0
botocore>=1.29.0
cryptography>=44.0.2